Rename pollUsers files to match kebab case naming scheme
